Scripted file to adjust SQL template to import data sets to new tables more automatically.

+++ data/world_bank/create_table_gdp_per_capita.sql
@@ -3,11 +3,11 @@ CREATE TABLE worldbank_gdp_per_capita (
     id SERIAL,
     country VARCHAR(64),
     country_code VARCHAR(4),
-    year_2010 SMALLINT,
-    year_2012 SMALLINT,
-    year_2014 SMALLINT,
-    year_2016 SMALLINT,
-    year_2018 SMALLINT,
-    year_2020 SMALLINT,
+    year_2010 INT,
+    year_2012 INT,
+    year_2014 INT,
+    year_2016 INT,
+    year_2018 INT,
+    year_2020 INT,
     PRIMARY KEY (id)
 );
